refactor(types): promote protocol types to agentex.protocol.*
Moves the JSON-RPC envelope types and ACP method-param types out of the
hand-authored ADK overlay (agentex.lib.types.*) into a new canonical
location at agentex.protocol.*. Back-compat shims at the old paths
re-export the same classes, so existing imports continue to work.

Motivation:
The existing layout puts protocol-shape types under agentex.lib.types.
That means a REST-only consumer who just wants typed JSON-RPC envelopes
(e.g. egp-api-backend, which today hand-rolls a ~600-line gateway
constructing `{"jsonrpc": "2.0", "method": "task/create", ...}` dicts by
hand) can't import the types without pulling in the full heavy ADK
runtime (~31 deps including temporalio, fastapi, claude-agent-sdk, etc.).
This is the first step toward letting that consumer drop hand-rolled
dict literals in favor of typed CreateTaskParams / SendMessageParams /
SendEventParams / JSONRPCRequest / JSONRPCResponse models, ahead of the
forthcoming slim-package split (a separate PR).

Scope:
- Move src/agentex/lib/types/acp.py → src/agentex/protocol/acp.py
- Move src/agentex/lib/types/json_rpc.py → src/agentex/protocol/json_rpc.py
- json_rpc.py: switch from `agentex.lib.utils.model_utils.BaseModel`
  (which transitively imports pyyaml) to `pydantic.BaseModel` directly.
  These classes don't use any model_utils extensions (from_yaml, to_json,
  populate_by_name) so the swap is behavior-preserving.
- Add re-export shims at the old paths so existing
  `from agentex.lib.types.{acp,json_rpc} import ...` imports continue
  to work. Identity check confirms the shimmed and canonical classes
  are the same objects.
- Update all internal usages within agentex.lib.* to import from the
  canonical path (9 files in src/agentex/lib + 1 in tests/).

Files moved use only pydantic + agentex.types.{Task,Agent,Event,
TaskMessageContent} — all slim-safe deps. Other agentex.lib.types.*
modules with heavier deps (fastacp pulls temporal, converters pulls
openai-agents) are left in place.

Verified locally:
- ruff check . → All checks passed
- Wheel build → ships both agentex/protocol/* and shims at
  agentex/lib/types/{acp,json_rpc}.py
- Import test from a fresh editable install: both canonical and shim
  paths resolve to the same class objects (identity-checked)

Zero install-time impact for existing consumers — same import paths
keep working.

Tracking: AGX1-292 (the slim/heavy split that motivates this refactor
lands as a follow-up PR).
